requirements.txt
pandas==3.0.2
numpy==2.4.4
SQLAlchemy==2.0.49
PyMySQL==1.1.2
python-dotenv==1.1.1
ipykernel==7.2.0
ipython==9.10.1
jupyter_client==8.8.0
jupyter_core==5.9.1
